Phase 1: extract alert report and recommendations into submodules
